+++ packages/Peptides/src/utils/constants.ts
@@ -29,6 +29,7 @@ export enum TAGS {
   POSITION = 'pos',
   SEPARATOR = 'separator',
   SELECTION = 'selection',
+  MUTATION_CLIFFS_SELECTION = 'mutationCliffsSelection',
   ALPHABET = 'alphabet',
   FILTER = 'filter',
   INVARIANT_MAP_SELECTION = 'invariantMapSelection',