+++ packages/Peptides/src/utils/types.ts
@@ -8,11 +8,9 @@ export type RawData = Int32Array | Uint32Array | Float32Array | Float64Array;
 export type UTypedArray = Uint8Array | Uint16Array | Uint32Array;
 //AAR: (Position: (index: indexList))
 export type MutationCliffs = Map<string, Map<string, Map<number, number[] | UTypedArray>>>;
-export type PositionToAARList = {[postiton: string]: string[]};
-
-export type MonomerSelectionStats = {[position: string]: {[monomer: string]: number}};
-export type MonomerPositionPair = {monomer: string, position: string};
-export type Cluster = {name: string, type: ClusterType};
+export type Selection = {[positionOrClusterType: string | ClusterType]: string[]};
+export type SelectionItem = {positionOrClusterType: string | ClusterType, monomerOrCluster: string};
+export type SelectionStats = {[positionOrClusterType: string | ClusterType]: { [monomerOrCluster: string]: number}};
 
 export type PeptidesSettings = {
   sequenceColumnName?: string,
@@ -51,4 +49,4 @@ export type StatsInfo = {
 
 export type RawColumn = {name: string, rawData: RawData, cat?: string[]};
 
-export type CustomClusters = {[clusterName: string]: number[] | Int32Array};
+export type SelectionOptions = {shiftPressed: boolean, ctrlPressed: boolean};